refactor(xtask): decompose benchmark.rs into submodules (SPL-22 #2450)
Split the 1604-line monolithic benchmark command into a focused module
tree under `xtask/src/commands/benchmark/`:

- `mod.rs` (164) - public surface and main dispatch
- `config.rs` (175) - `BenchmarkOptions`, constants, mirror catalogue
- `runner.rs` (402) - timing loop, daemon lifecycle, version builds, stat parsing
- `report.rs` (294) - text-table and JSON formatters
- `loopback.rs` (495) - per-version daemon orchestration and data generation
- `tests.rs` (131) - existing unit tests, re-pointed at submodule paths

Public API unchanged: `benchmark::execute` and `benchmark::BenchmarkOptions`
remain the only items consumed by `main.rs`. All cross-submodule helpers use
`pub(super)` so nothing leaks beyond the `benchmark` module.
